Refactor - organise files, remove unused code and resolve inspections
