+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/client/PrisonerSearchClient.kt
@@ -2,6 +2,7 @@ package uk.gov.justice.digital.hmpps.personrecord.client
 
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
+import org.springframework.web.reactive.function.client.bodyToMono
 import uk.gov.justice.digital.hmpps.personrecord.client.model.prisoner.Prisoner
 import uk.gov.justice.digital.hmpps.personrecord.model.person.Person
 import uk.gov.justice.digital.hmpps.personrecord.service.queue.discardNotFoundException
@@ -13,7 +14,7 @@ class PrisonerSearchClient(private val prisonerSearchWebClient: WebClient) {
     .get()
     .uri("/prisoner/{id}", prisonNumber)
     .retrieve()
-    .bodyToMono(Prisoner::class.java)
+    .bodyToMono<Prisoner>()
     .discardNotFoundException()
     .block()?.let { return Person.from(it) }
 }
