--- src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/client/model/PrisonNumbers.kt
@@ -1,14 +1,0 @@
-package uk.gov.justice.digital.hmpps.personrecord.client.model
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-import com.fasterxml.jackson.annotation.JsonProperty
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class PrisonNumbers(
-
-  val totalPages: Int,
-  val first: Boolean,
-  val last: Boolean,
-  @JsonProperty("content")
-  val numbers: List<String>,
-)
