+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/client/model/prisoner/PhoneNumber.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.personrecord.client.model
+package uk.gov.justice.digital.hmpps.personrecord.client.model.prisoner
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
