+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/client/model/prisoner/Prisoner.kt
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
 import jakarta.validation.constraints.NotBlank
 import tools.jackson.databind.annotation.JsonDeserialize
-import uk.gov.justice.digital.hmpps.personrecord.client.model.PhoneNumber
 import uk.gov.justice.digital.hmpps.personrecord.model.identifiers.CROIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.model.identifiers.CROIdentifierDeserializer
 import uk.gov.justice.digital.hmpps.personrecord.model.identifiers.PNCIdentifier
